[skip ci] mvSliderItems.h and mvAboutWindow.h splited into *.hpp and *.cpp files (#391)
* mvSliderItems.h and mvAboutWindow.h splited into *.hpp and *.cpp files, please, check, is there ok with #include directives?

* mvSliderItems.h and mvAboutWindow.h splited into *.h and *.cpp files, please, check, is there ok with #include directives?
